Split generic type reference walker
