+++ docs/PHASE_PLAN.md
@@ -337,6 +337,9 @@ checked-in docs, tests, and commits only.
   conversion.
 - Pattern checking now keeps enum and bool match validation in a focused child
   module, leaving core pattern binding and lowering helpers smaller.
+- Generic type reference validation now keeps expression and statement walking
+  in a focused child module, leaving generic type argument arity and bound
+  checks smaller.
 - Generic diagnostic integration coverage now keeps nested/function/container
   annotation arity tests in a focused module, separate from direct generic
   call, method, local, and declaration annotation arity cases.
