+++ src/typechecker/generic_type_reference_walker.rs
@@ -1,5 +1,7 @@
 use super::*;
 
+mod expressions;
+
 impl TypeChecker {
     pub(super) fn validate_generic_type_ref_bounds(
         &mut self,
@@ -223,205 +225,5 @@ impl TypeChecker {
         self.structs.contains_key(name)
             || self.enums.contains_key(name)
             || self.imports.contains_key(name)
-    }
-
-    pub(super) fn validate_generic_expr_type_references(
-        &mut self,
-        expr: &Expression,
-        scoped_type_params: &HashSet<String>,
-    ) {
-        match expr {
-            Expression::FunctionCall {
-                type_args,
-                args,
-                span,
-                ..
-            } => {
-                self.validate_generic_type_arg_refs(type_args, scoped_type_params, *span);
-                for arg in args {
-                    self.validate_generic_expr_type_references(arg, scoped_type_params);
-                }
-            }
-            Expression::MethodCall {
-                receiver,
-                type_args,
-                args,
-                span,
-                ..
-            } => {
-                self.validate_generic_expr_type_references(receiver, scoped_type_params);
-                self.validate_generic_type_arg_refs(type_args, scoped_type_params, *span);
-                for arg in args {
-                    self.validate_generic_expr_type_references(arg, scoped_type_params);
-                }
-            }
-            Expression::BinaryOp { left, right, .. } => {
-                self.validate_generic_expr_type_references(left, scoped_type_params);
-                self.validate_generic_expr_type_references(right, scoped_type_params);
-            }
-            Expression::UnaryOp { operand, .. } => {
-                self.validate_generic_expr_type_references(operand, scoped_type_params);
-            }
-            Expression::MemberAccess { object, .. } => {
-                self.validate_generic_expr_type_references(object, scoped_type_params);
-            }
-            Expression::IndexAccess { object, index, .. } => {
-                self.validate_generic_expr_type_references(object, scoped_type_params);
-                self.validate_generic_expr_type_references(index, scoped_type_params);
-            }
-            Expression::StructLiteral {
-                type_args,
-                fields,
-                span,
-                ..
-            } => {
-                self.validate_generic_type_arg_refs(type_args, scoped_type_params, *span);
-                for (_, value) in fields {
-                    self.validate_generic_expr_type_references(value, scoped_type_params);
-                }
-            }
-            Expression::EnumVariant {
-                type_args,
-                payload,
-                span,
-                ..
-            } => {
-                self.validate_generic_type_arg_refs(type_args, scoped_type_params, *span);
-                if let Some(payload) = payload {
-                    self.validate_generic_expr_type_references(payload, scoped_type_params);
-                }
-            }
-            Expression::ArrayLiteral { elements, .. } => {
-                for element in elements {
-                    self.validate_generic_expr_type_references(element, scoped_type_params);
-                }
-            }
-            Expression::Match {
-                scrutinee, arms, ..
-            } => {
-                self.validate_generic_expr_type_references(scrutinee, scoped_type_params);
-                for arm in arms {
-                    if let Some(guard) = &arm.guard {
-                        self.validate_generic_expr_type_references(guard, scoped_type_params);
-                    }
-                    self.validate_generic_expr_type_references(&arm.body, scoped_type_params);
-                }
-            }
-            Expression::WhileLoop {
-                condition, body, ..
-            } => {
-                self.validate_generic_expr_type_references(condition, scoped_type_params);
-                self.validate_generic_expr_type_references(body, scoped_type_params);
-            }
-            Expression::Loop { body, .. } => {
-                self.validate_generic_expr_type_references(body, scoped_type_params);
-            }
-            Expression::If {
-                condition,
-                then_body,
-                else_body,
-                ..
-            } => {
-                self.validate_generic_expr_type_references(condition, scoped_type_params);
-                self.validate_generic_expr_type_references(then_body, scoped_type_params);
-                if let Some(else_body) = else_body {
-                    self.validate_generic_expr_type_references(else_body, scoped_type_params);
-                }
-            }
-            Expression::Block {
-                statements, expr, ..
-            } => {
-                for statement in statements {
-                    self.validate_generic_statement_type_references(statement, scoped_type_params);
-                }
-                if let Some(expr) = expr {
-                    self.validate_generic_expr_type_references(expr, scoped_type_params);
-                }
-            }
-            Expression::Return { value, .. } => {
-                if let Some(value) = value {
-                    self.validate_generic_expr_type_references(value, scoped_type_params);
-                }
-            }
-            Expression::Closure {
-                params,
-                return_type,
-                body,
-                span,
-            } => {
-                for param in params {
-                    self.validate_generic_type_ref_bounds(
-                        &param.ty,
-                        scoped_type_params,
-                        param.span,
-                    );
-                }
-                if let Some(return_type) = return_type {
-                    self.validate_generic_type_ref_bounds(return_type, scoped_type_params, *span);
-                }
-                self.validate_generic_expr_type_references(body, scoped_type_params);
-            }
-            Expression::Cast {
-                expr,
-                target_type,
-                span,
-            } => {
-                self.validate_generic_expr_type_references(expr, scoped_type_params);
-                self.validate_generic_type_ref_bounds(target_type, scoped_type_params, *span);
-            }
-            Expression::StringInterpolation { parts, .. } => {
-                for part in parts {
-                    if let ast::StringPart::Expr(expr) = part {
-                        self.validate_generic_expr_type_references(expr, scoped_type_params);
-                    }
-                }
-            }
-            Expression::Range { start, end, .. } => {
-                self.validate_generic_expr_type_references(start, scoped_type_params);
-                self.validate_generic_expr_type_references(end, scoped_type_params);
-            }
-            Expression::Defer { expr, .. } => {
-                self.validate_generic_expr_type_references(expr, scoped_type_params);
-            }
-            Expression::IntLiteral { .. }
-            | Expression::FloatLiteral { .. }
-            | Expression::StringLiteral { .. }
-            | Expression::BoolLiteral { .. }
-            | Expression::CharLiteral { .. }
-            | Expression::Identifier { .. }
-            | Expression::Break { .. }
-            | Expression::Continue { .. }
-            | Expression::LoopControl { .. }
-            | Expression::Error { .. } => {}
-        }
-    }
-
-    fn validate_generic_statement_type_references(
-        &mut self,
-        statement: &ast::Statement,
-        scoped_type_params: &HashSet<String>,
-    ) {
-        match statement {
-            ast::Statement::VarDecl {
-                ty, value, span, ..
-            } => {
-                if let Some(ty) = ty {
-                    self.validate_generic_type_ref_bounds(ty, scoped_type_params, *span);
-                }
-                self.validate_generic_expr_type_references(value, scoped_type_params);
-            }
-            ast::Statement::Assignment { target, value, .. } => {
-                self.validate_generic_expr_type_references(target, scoped_type_params);
-                self.validate_generic_expr_type_references(value, scoped_type_params);
-            }
-            ast::Statement::Expression { expr, .. } => {
-                self.validate_generic_expr_type_references(expr, scoped_type_params);
-            }
-            ast::Statement::Block { stmts, .. } => {
-                for statement in stmts {
-                    self.validate_generic_statement_type_references(statement, scoped_type_params);
-                }
-            }
-        }
     }
 }
